Move core functionalities from common to src
