+++ resources/links/LinksGetterBySearch.py
@@ -3,8 +3,7 @@ from flask import jsonify, make_response
 from flask_jwt_extended import jwt_required, get_jwt_identity
 from flasgger import swag_from
 
-from common.DbHandler import DbHandler
-
+from src.DbHandler import DbHandler
 import json
 
 # Load config file
