+++ resources/user/Publicity.py
@@ -3,7 +3,7 @@ from flask import jsonify, make_response
 from flask_jwt_extended import jwt_required, get_jwt_identity
 from flasgger import swag_from
 
-from common.DbHandler import DbHandler
+from src.DbHandler import DbHandler
 
 
 class Publicity(Resource):
